[api] Make Typescript works !
Define a tsconfig.json project configuration file.

Module had to be set to "es6" (instead of documentation default "es5") because of the *manual generic inheritance mechanism* for Sequelize models.

We use 2 dependencies, cf. https://vccolombo.github.io/blog/tsc-how-to-copy-non-typescript-files-when-building/:
- rimraf, "the Unix `rm -rf` command for Node.js"
- copyfiles

We add multiples NPM tasks in order to transpile the source files and build the `dist` output.